% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate a report for the accelerometer data}
\usage{
generate_report(file_path, participant_id, parent_name, child_name)
}
\arguments{
\item{file_path}{The path to the accelerometer file.
Needs to be the same path as the \code{process_file} function.}

\item{participant_id}{The participant ID}

\item{parent_name}{Name of the parent}

\item{child_name}{Name of the child}
}
\value{
Path to the generated report
}
\description{
Generate a report for the accelerometer data
}
