% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_file.R
\name{process_file}
\alias{process_file}
\title{Process an accelerometer file}
\usage{
process_file(file_path, participant_id, overwrite = FALSE)
}
\arguments{
\item{file_path}{The path to the accelerometer file}

\item{participant_id}{The participant ID}

\item{overwrite}{Overwrite existing files? Set to \code{TRUE} to rerun processing.}
}
\value{
Nothing. The function writes files to the \code{file_path} directory.
}
\description{
Process an accelerometer file
}
