% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_report.R
\name{upload_report}
\alias{upload_report}
\title{Generate a report for the accelerometer data}
\usage{
upload_report(
  report_path,
  participant_id,
  timepoint,
  redcap_uri = Sys.getenv("REDCAP_URI"),
  redcap_key = Sys.getenv("REDCAP_KEY")
)
}
\arguments{
\item{report_path}{The path to the report file.}

\item{participant_id}{The participant ID}

\item{timepoint}{The time point to upload to. Must be one of "baseline",
"12m", or "24m"}

\item{redcap_uri}{REDCap endpoint. Defaults to the value of the environment
variable \code{REDCAP_URI}}

\item{redcap_key}{REDCap API key. Defaults to the value of the environment
variable \code{REDCAP_KEY}}
}
\value{
The output of \code{\link[redcapAPI:importRecords]{redcapAPI::importRecords()}}.
}
\description{
Generate a report for the accelerometer data
}
